Revert "Merge branch 'EliO/UpgradeTaproot' into 'develop'"
This reverts commit 483abb5c3541d2b98e577ad9462451c42f0beb2e, reversing
changes made to 08435ed124effcb46d0c8a477a4b2619326dec7f.

+++ aruw-mcb-project/src/aruwsrc/control/auto-aim/auto_aim_launch_timer.cpp
@@ -64,8 +64,7 @@ AutoAimLaunchTimer::LaunchInclination AutoAimLaunchTimer::getCurrentLaunchInclin
         return LaunchInclination::GATED_DENY;
     }
 
-    uint32_t timeOfFlightMicros =
-        static_cast<uint32_t>(static_cast<double>(timeOfFlightSeconds) * 1e6);
+    uint32_t timeOfFlightMicros = timeOfFlightSeconds * 1e6;
     uint32_t now = tap::arch::clock::getTimeMicroseconds();
     uint32_t projectedHitTime = now + this->agitatorTypicalDelayMicroseconds + timeOfFlightMicros;
